+++ Advent of Code/Day 07/Tests/Day7_Examples_Part2.cs
@@ -16,7 +16,7 @@ namespace Advent_of_Code.Day_07.Tests
             feedbackAmplifierTuner.Run();
 
             int expectedSignal = 139629729;
-            List<int> expectedTune = new List<int>() { 9, 8, 7, 6, 5 };
+            List<long> expectedTune = new List<long>() { 9, 8, 7, 6, 5 };
             Assert.Equal(expectedSignal, feedbackAmplifierTuner.HighestOutput);
             Assert.Equal(expectedTune, feedbackAmplifierTuner.OptimalTune);
         }
@@ -29,7 +29,7 @@ namespace Advent_of_Code.Day_07.Tests
             feedbackAmplifierTuner.Run();
 
             int expectedSignal = 18216;
-            List<int> expectedTune = new List<int>() { 9, 7, 8, 5, 6 };
+            List<long> expectedTune = new List<long>() { 9, 7, 8, 5, 6 };
             Assert.Equal(expectedSignal, feedbackAmplifierTuner.HighestOutput);
             Assert.Equal(expectedTune, feedbackAmplifierTuner.OptimalTune);
         }
