+++ Advent of Code/SharedLibrary/IntcodeVirtualMachine/Input-OutputProviders/ConsoleOutputProvider.cs
@@ -1,12 +1,10 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Text;
 
 namespace Advent_of_Code.SharedLibrary.IntcodeVirtualMachine.Input_OutputProviders
 {
     class ConsoleOutputProvider : IOutputProvider
     {
-        public void SendOutput(int output)
+        public void SendOutput(long output)
         {
             Console.WriteLine($"Output: {output}");
         }
